:sparkles: fix reminder settings widget and use it
