--- lib/settings/list/reminder_time.dart
@@ -1,37 +1,0 @@
-import 'package:Pilll/store/setting.dart';
-import 'package:Pilll/theme/color.dart';
-import 'package:Pilll/theme/text_color.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/src/widgets/framework.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:hooks_riverpod/all.dart';
-
-class ReminderTime extends HookWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: PilllColors.background,
-      appBar: AppBar(
-        leading: IconButton(
-          icon: Icon(Icons.arrow_back, color: Colors.black),
-          onPressed: () => Navigator.of(context).pop(),
-        ),
-        title: Text(
-          "通知時間",
-          style: TextStyle(color: TextColor.black),
-        ),
-        backgroundColor: PilllColors.background,
-      ),
-      body: Container(
-        child: ListView(
-          children: _components(context),
-        ),
-      ),
-    );
-  }
-
-  List<Widget> _components(BuildContext context) {
-    final state = useProvider(settingStoreProvider.state);
-    return [];
-  }
-}
