+++ lib/settings/list/settings.dart
@@ -9,6 +9,7 @@ import 'package:Pilll/model/user.dart';
 import 'package:Pilll/provider/auth.dart';
 import 'package:Pilll/settings/list/model.dart';
 import 'package:Pilll/settings/list/modifing_pill_number.dart';
+import 'package:Pilll/settings/list/reminder_times.dart';
 import 'package:Pilll/store/pill_sheet.dart';
 import 'package:Pilll/store/setting.dart';
 import 'package:Pilll/style/button.dart';
@@ -193,20 +194,10 @@ class Settings extends HookWidget {
             content: DateTimeFormatter.militaryTime(
                 settingState.entity.reminderDateTime()),
             onTap: () {
-              showModalBottomSheet(
-                context: context,
-                builder: (BuildContext context) {
-                  return DateTimePicker(
-                    initialDateTime: settingState.entity.reminderDateTime(),
-                    done: (dateTime) {
-                      Navigator.pop(context);
-                      settingStore.modifyReminderTime(
-                        ReminderTime(
-                            hour: dateTime.hour, minute: dateTime.minute),
-                      );
-                    },
-                  );
-                },
+              Navigator.of(context).push(
+                MaterialPageRoute(builder: (BuildContext context) {
+                  return ReminderTimes();
+                }),
               );
             },
           ),
